Boundary test now passes.

+++ src/Ravl2/Image/Segmentation/Segmentation.cc
@@ -18,12 +18,12 @@
 namespace Ravl2
 {
 
-  SegmentationBodyC::SegmentationBodyC(const Array<unsigned , 2> &nsegmap)
+  SegmentationBodyC::SegmentationBodyC(const Array<int , 2> &nsegmap)
       : segmap(nsegmap.range()),
         labels(0)
   {
     for(auto it = begin(segmap,nsegmap);it.valid();++it) {
-      it.data<0>() = it.data<1>() >= 0 ? it.data<1>() : 0;
+      it.data<0>() = unsigned(it.data<1>() >= 0 ? it.data<1>() : 0);
       if(it.data<0>() > labels)
         labels = it.data<0>();
     }
